refactor : Score를 인터페이스와 추상클래스로 분리
