+++ src/main/java/bowling/domain/score/NormalScore.java
@@ -5,14 +5,9 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 
-public class NormalScore extends Score {
+public class NormalScore extends BaseScore {
 
     private final boolean isDone;
-
-    private NormalScore(int score, boolean isDone) {
-        super(score, -1);
-        this.isDone = isDone;
-    }
 
     private NormalScore(int first, int second, boolean isDone) {
         super(first, second);
@@ -22,36 +17,29 @@ public class NormalScore extends Score {
 
     public static NormalScore first(int score) {
         validateScore(score);
-        return new NormalScore(score, false);
+        return new NormalScore(score, -1, false);
     }
 
     public NormalScore second(int score) {
         validateScore(score);
         validateCombinedScores(score);
-        return new NormalScore(this.first, score, true);
+        return new NormalScore(getFirst(), score, true);
     }
 
     public boolean isDone() {
         return isDone;
     }
 
-    public int sum() {
-        if (isStrike()) {
-            return MAX;
-        }
-        return first + second;
-    }
-
     @Override
     protected void validateCombinedScores(int score) {
-        if (this.first + score > MAX) {
+        if (getFirst() + score > MAX) {
             throw new IllegalArgumentException("1차시도와 2차시도의 합계는 10점을 넘을 수 없습니다.");
         }
     }
 
     @Override
     public List<Integer> getAll() {
-        return Collections.unmodifiableList(Arrays.asList(first, second));
+        return Collections.unmodifiableList(Arrays.asList(getFirst(), getSecond()));
     }
 
     @Override
@@ -59,11 +47,11 @@ public class NormalScore extends Score {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         NormalScore that = (NormalScore) o;
-        return first == that.first && second == that.second && isDone == that.isDone;
+        return isDone == that.isDone;
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(first, second, isDone);
+        return Objects.hash(isDone);
     }
 }
